feat: improve rendering across various media queries
Enhance the rendering experience on different devices, with a focus on
better support for smaller devices with narrower client widths.

Additionally, custom hooks from the misc file have been relocated to the
UI custom hooks file for better organization and consistency.
